Modify logic of SettingWrapper.dart
